Split test_invalids into syntax and exceptions cases.
