+++ tests/parser/exceptions/test_invalid_payable.py
@@ -1,0 +1,39 @@
+import pytest
+from pytest import raises
+
+from viper import compiler
+from viper.exceptions import NonPayableViolationException
+
+
+fail_list = [
+    """
+def foo():
+    x = msg.value
+"""
+]
+
+
+@pytest.mark.parametrize('bad_code', fail_list)
+def test_variable_decleration_exception(bad_code):
+        with raises(NonPayableViolationException):
+            compiler.compile(bad_code)
+
+
+valid_list = [
+    """
+x: num
+@payable
+def foo() -> num:
+    self.x = 5
+    """,
+    """
+@payable
+def foo():
+    x = msg.value
+    """
+]
+
+
+@pytest.mark.parametrize('good_code', valid_list)
+def test_block_success(good_code):
+    assert compiler.compile(good_code) is not None
