+++ tests/parser/syntax/test_as_wei.py
@@ -1,0 +1,52 @@
+import pytest
+from pytest import raises
+
+from viper import compiler
+from viper.exceptions import TypeMismatchException
+
+
+fail_list = [
+    """
+def foo():
+    x = as_wei_value(5, 'szabo')
+    """,
+    """
+def foo() -> num(wei):
+    x = 45
+    return x.balance
+    """
+]
+
+
+@pytest.mark.parametrize('bad_code', fail_list)
+def test_as_wei_fail(bad_code):
+
+    with raises(TypeMismatchException):
+        compiler.compile(bad_code)
+
+
+valid_list = [
+    """
+def foo():
+    x = as_wei_value(5, finney) + as_wei_value(2, babbage) + as_wei_value(8, shannon)
+    """,
+    """
+def foo():
+    z = 2 + 3
+    x = as_wei_value(2 + 3, finney)
+    """,
+    """
+def foo():
+    x = as_wei_value(5.182, ada)
+    """,
+    """
+def foo() -> num(wei):
+    x = 0x1234567890123456789012345678901234567890
+    return x.balance
+    """
+]
+
+
+@pytest.mark.parametrize('good_code', valid_list)
+def test_as_wei_success(good_code):
+    assert compiler.compile(good_code) is not None
