+++ tests/parser/syntax/test_code_size.py
@@ -1,0 +1,40 @@
+import pytest
+from pytest import raises
+
+from viper import compiler
+from viper.exceptions import TypeMismatchException
+
+
+fail_list = [
+    """
+def foo() -> num:
+    x = 45
+    return x.codesize
+    """,
+    """
+def foo() -> num(wei):
+    x = 0x1234567890123456789012345678901234567890
+    return x.codesize
+    """
+]
+
+
+@pytest.mark.parametrize('bad_code', fail_list)
+def test_block_fail(bad_code):
+
+    with raises(TypeMismatchException):
+        compiler.compile(bad_code)
+
+
+valid_list = [
+    """
+def foo() -> num:
+    x = 0x1234567890123456789012345678901234567890
+    return x.codesize
+    """
+]
+
+
+@pytest.mark.parametrize('good_code', valid_list)
+def test_block_success(good_code):
+    assert compiler.compile(good_code) is not None
