+++ tests/parser/syntax/test_for_range.py
@@ -1,0 +1,27 @@
+import pytest
+
+from viper import compiler
+
+valid_list = [
+    """
+def foo():
+    for i in range(10):
+        pass
+    """,
+    """
+def foo():
+    for i in range(10, 20):
+        pass
+    """,
+    """
+def foo():
+    x = 5
+    for i in range(x, x + 10):
+        pass
+    """
+]
+
+
+@pytest.mark.parametrize('good_code', valid_list)
+def test_range_success(good_code):
+    assert compiler.compile(good_code) is not None
